refactor: 移动 Gateway Model 单独到 gateway 文件中
