--- apps/assets/models/asset/host/__init__.py
@@ -1,2 +1,0 @@
-from .host import *
-from .gateway import *
